drawing functions are transferred to the app.js file
